Move graph-spec modules into environment_spec package
Relocate the four arena_env_graph_* modules (spec, types, and the two
conversion utils) out of environments/ into a new environment_spec/
package to separate the data-only graph schema from runtime env code,
and update all import sites accordingly.
