+++ isaaclab_arena/agentic_environment_generation/spec_validation.py
@@ -13,7 +13,7 @@ from typing import Any
 from pydantic import ValidationError
 
 from isaaclab_arena.assets.registries import TaskRegistry
-from isaaclab_arena.environments.arena_env_graph_spec import ArenaEnvGraphSpec
+from isaaclab_arena.environment_spec.arena_env_graph_spec import ArenaEnvGraphSpec
 
 
 def required_task_init_param_names(task_cls: type) -> list[str]:
