+++ isaaclab_arena_examples/agentic_environment_generation/review_gui/editor_panel.py
@@ -14,7 +14,7 @@ import streamlit as st
 from streamlit_ace import st_ace
 
 from isaaclab_arena.agentic_environment_generation.spec_io import env_graph_spec_path, write_env_graph_spec
-from isaaclab_arena.environments.arena_env_graph_spec import ArenaEnvGraphSpec
+from isaaclab_arena.environment_spec.arena_env_graph_spec import ArenaEnvGraphSpec
 
 
 @dataclass
