+++ isaaclab_arena_examples/agentic_environment_generation/review_gui/simapp/sim_preview.py
@@ -16,8 +16,8 @@ from typing import Any
 
 from isaaclab.envs.common import ViewerCfg
 
+from isaaclab_arena.environment_spec.arena_env_graph_spec import ArenaEnvGraphSpec
 from isaaclab_arena.environments.arena_env_builder_cfg import ArenaEnvBuilderCfg
-from isaaclab_arena.environments.arena_env_graph_spec import ArenaEnvGraphSpec
 from isaaclab_arena.utils.isaaclab_utils.simulation_app import (
     collect_garbage_and_clear_cuda_cache,
     teardown_simulation_app,
