+++ isaaclab_arena_examples/agentic_environment_generation/review_gui/simapp/thumbnail_capture.py
@@ -14,7 +14,7 @@ import omni.usd
 from omni.kit.viewport.utility import frame_viewport_prims, get_active_viewport
 from pxr import Gf, Sdf, UsdGeom, UsdLux
 
-from isaaclab_arena.environments.arena_env_graph_spec import ArenaEnvGraphSpec
+from isaaclab_arena.environment_spec.arena_env_graph_spec import ArenaEnvGraphSpec
 from isaaclab_arena_examples.agentic_environment_generation.review_gui.simapp.asset_usd import (
     AabbDimensionsM,
     resolve_node_aabb_dimensions_m,
